[BUGFIX] Test for valid JSON in `JsonResponse::supports()`
Since the cPanel API does not send the correct `Content-Type`
response header, we can not safely rely on this information.
Therefore, the response body is now JSON-decoded. If this is
successful, the response is considered "supported". Note
that the JSON decode is meant as "fallback". The response
headers are checked first.